requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mp3play==0.1.15
psutil==6.0.0
pycparser==2.22
pydub==0.25.1
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
simple-websocket==1.0.0
srt==3.5.3
tqdm==4.66.4
urllib3==2.2.2
vosk==0.3.44
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0

black~=24.8.0
pyjokes~=0.8.3
websocket-client~=1.8.0
comtypes~=1.4.9
pycaw~=20240210